+++ __test__/backend.test.js
@@ -12,9 +12,7 @@ import {
   CannedResponse,
   InteractionStep,
   UserOrganization,
-  Tag,
-  Message,
-  TagCampaignContact
+  Message
 } from "../src/server/models/";
 import { resolvers as campaignResolvers } from "../src/server/api/campaign";
 import {
@@ -821,188 +819,6 @@ describe("editUser mutation", () => {
       email: userData.email,
       cell: userData.cell,
       id: testTexter.id
-    });
-  });
-});
-
-describe("A tag table", () => {
-  let userTest;
-  let inviteTest;
-  let organizationTest;
-  let campaignTest;
-  let contactTest;
-  let tagFields;
-  let campaignContactFromDB;
-  beforeEach(async () => {
-    await setupTest();
-
-    // Creating the instances needed to create a new tag and update its content
-    userTest = await createUser();
-    inviteTest = await createInvite();
-    organizationTest = await createOrganization(
-      userTest,
-      "Test org for tag",
-      inviteTest.data.createInvite.id,
-      inviteTest.data.createInvite.id
-    );
-    campaignTest = await createCampaign(
-      userTest,
-      "A campaign test for tags",
-      "A campaign test for tags",
-      organizationTest.data.createOrganization.id
-    );
-    contactTest = await createContact(campaignTest.data.createCampaign.id);
-
-    // createCampaign feeds the campaign_contact table, so we can use it from de db to create the tag content
-    campaignContactFromDB = await r.knex("campaign_contact").where({
-      campaign_id: contactTest.campaign_id,
-      first_name: contactTest.first_name,
-      last_name: contactTest.last_name
-    });
-
-    // a tag that will be used accross most of the tets
-    tagFields = {
-      name: "A new tag",
-      description: "This is a description for a new tag",
-      organization_id: organizationTest.data.createOrganization.id,
-      group: "A group for my tag"
-    };
-    await new Tag(tagFields).save();
-  }, global.DATABASE_SETUP_TEARDOWN_TIMEOUT);
-
-  afterEach(async () => {
-    await cleanupTest();
-    if (r.redis) r.redis.flushdb();
-  }, global.DATABASE_SETUP_TEARDOWN_TIMEOUT);
-
-  it("holds all tags with name, description, andr group, and tags belong to organizations", async () => {
-    // find from DB the Tag created
-    const [tagFromDB] = await Tag.getAll();
-    // it should match the object that was used to create it
-    expect(tagFromDB).toMatchObject(tagFields);
-  });
-  it("name, description, and organization cannot be null. Group can be null", async () => {
-    // create new Tags, where not nullable fields receive null as their value
-    const incompletedTags = [
-      { ...tagFields, ...{ name: null } },
-      { ...tagFields, ...{ description: null } },
-      { ...tagFields, ...{ organization_id: null } }
-    ];
-    const tagsDBInstance = [];
-
-    incompletedTags.forEach(tag => {
-      tagsDBInstance.push(new Tag(tag));
-    });
-
-    // For some reason, I couldn't make .toThrow() to work. So catching all errors and comparing with the incompletedTags array
-    const catchErrors = [];
-
-    for (let i = 0; i < tagsDBInstance.length; i++) {
-      try {
-        await tagsDBInstance[i].save();
-      } catch (e) {
-        catchErrors.push(true);
-      }
-    }
-    expect(catchErrors.length).toBe(incompletedTags.length);
-  });
-  it("Group can be null", async () => {
-    const tagWithoutGroup = { ...tagFields, ...{ group: null } };
-
-    await new Tag(tagWithoutGroup).save();
-    const [tagCreated] = await Tag.getAll();
-    expect(tagCreated.group).toBe(null);
-  });
-  it("has a default field 'is_deleted' as false, but can be updated to true", async () => {
-    const [tagFromDB] = await Tag.getAll();
-    expect(tagFromDB.isDeleted).toBeFalsy();
-    await r
-      .knex("tag")
-      .where({ id: tagFromDB.id })
-      .update({ is_deleted: true });
-    const [tagFromDBAfterUpdate] = await Tag.getAll();
-    expect(tagFromDBAfterUpdate).toBeTruthy();
-  });
-  describe("and a tag", () => {
-    let tagTest;
-    let tagCampaignContent;
-
-    beforeEach(async () => {
-      await setupTest();
-
-      // Get tag needed to test from the DB
-      [tagTest] = await Tag.getAll();
-
-      tagCampaignContent = {
-        value: "Tagged a contact",
-        tag_id: tagTest.id,
-        campaign_contact_id: campaignContactFromDB[0].id
-      };
-
-      // Tag a contact! used accross the tests below
-      await new TagCampaignContact(tagCampaignContent).save();
-    }, global.DATABASE_SETUP_TEARDOWN_TIMEOUT);
-
-    afterEach(async () => {
-      await cleanupTest();
-      if (r.redis) r.redis.flushdb();
-    }, global.DATABASE_SETUP_TEARDOWN_TIMEOUT);
-
-    it("is applied to a contact, and associated to a campaign, and can hold a value that explains why that tag was applied to that message", async () => {
-      // Check on db if the tag content was created
-      const [tagContentFromDB] = await TagCampaignContact.getAll();
-      expect(tagContentFromDB).toMatchObject(tagCampaignContent);
-    });
-
-    it("one tag can be applied to two different contacts", async () => {
-      // Create a new contact..
-      const newContact = await new CampaignContact({
-        campaign_id: campaignTest.data.createCampaign.id,
-        first_name: "Dwight",
-        last_name: "Schrute",
-        cell: "5555555555",
-        zip: "18503"
-      }).save();
-      // get it from db
-      const queryTheNewContact = await r.knex("campaign_contact").where({
-        campaign_id: Number(newContact.campaign_id),
-        first_name: newContact.first_name,
-        last_name: newContact.last_name
-      });
-      // ...and aply the same tag applied to another contact
-      const newTagCampaignContact = {
-        value: "Tagged a different contact",
-        tag_id: tagTest.id,
-        campaign_contact_id: queryTheNewContact[0].id
-      };
-      await new TagCampaignContact(newTagCampaignContact).save();
-
-      const newTagCampaignContactFromDB = await TagCampaignContact.getAll();
-      expect(newTagCampaignContactFromDB[0].campaign_contact_id).not.toBe(
-        newTagCampaignContactFromDB[1].campaign_contact_id
-      );
-    });
-    it("is possible to tag one contact with two or more tags", async () => {
-      const newTagFields = {
-        name: "A different tag",
-        description: "This is not the same tag",
-        organization_id: organizationTest.data.createOrganization.id
-      };
-      // Create a new tag...
-      const newTag = await new Tag(newTagFields).save();
-      const newTagCampaignContent = {
-        value: "Testing tagging same message two different tags",
-        tag_id: newTag.id,
-        campaign_contact_id: campaignContactFromDB[0].id
-      };
-      // and tag a contact that already was tagged
-      await new TagCampaignContact(newTagCampaignContent).save();
-      const tagContentFromDB = await TagCampaignContact.getAll();
-      expect(tagContentFromDB[0].campaign_contact_id).toBe(
-        tagContentFromDB[1].campaign_contact_id
-      );
-      // just check it's not the same tag
-      expect(tagContentFromDB[0].tag_id).not.toBe(tagContentFromDB[1].tag_id);
     });
   });
 });
